+++ tests/test_disco.c
@@ -98,7 +98,7 @@ void test_NX() {
 
   // debug
   printf("sent %zu bytes\n", out_len);
-  for (int i = 0; i < out_len; i++) {
+  for (size_t i = 0; i < out_len; i++) {
     printf("%02x", out[i]);
   }
   printf("\n");
@@ -134,7 +134,7 @@ void test_NX() {
 
   // debug
   printf("sent %zu bytes:\n", out_len);
-  for (int i = 0; i < out_len; i++) {
+  for (size_t i = 0; i < out_len; i++) {
     printf("%02x", out[i]);
   }
   printf("\n");
@@ -215,7 +215,7 @@ void test_IK() {
 
   // debug
   printf("sent %zu bytes\n", out_len);
-  for (int i = 0; i < out_len; i++) {
+  for (size_t i = 0; i < out_len; i++) {
     printf("%02x", out[i]);
   }
   printf("\n");
@@ -251,7 +251,7 @@ void test_IK() {
 
   // debug
   printf("sent %zu bytes:\n", out_len);
-  for (int i = 0; i < out_len; i++) {
+  for (size_t i = 0; i < out_len; i++) {
     printf("%02x", out[i]);
   }
   printf("\n");
